Move `set_row`, `set_column` from matrix.rs to edition.rs.
